+++ crates/router/src/core/payments.rs
@@ -4044,6 +4044,16 @@ where
             )
             .await?,
         ));
+    operation
+        .to_domain()?
+        .populate_payment_data(
+            state,
+            payment_data,
+            merchant_context,
+            business_profile,
+            &connector,
+        )
+        .await?;
 
     let mut router_data = payment_data
         .construct_router_data(
@@ -8884,7 +8894,7 @@ pub trait OperationSessionGetters<F> {
     #[cfg(feature = "v2")]
     fn get_merchant_connector_details(
         &self,
-    ) -> Option<api_models::payments::MerchantConnectorDetails>;
+    ) -> Option<common_types::domain::MerchantConnectorAuthDetails>;
 
     fn get_connector_customer_id(&self) -> Option<String>;
     fn get_whole_connector_response(&self) -> Option<String>;
@@ -9395,7 +9405,7 @@ impl<F: Clone> OperationSessionGetters<F> for PaymentIntentData<F> {
 
     fn get_merchant_connector_details(
         &self,
-    ) -> Option<api_models::payments::MerchantConnectorDetails> {
+    ) -> Option<common_types::domain::MerchantConnectorAuthDetails> {
         todo!()
     }
 
@@ -9591,7 +9601,7 @@ impl<F: Clone> OperationSessionGetters<F> for PaymentConfirmData<F> {
 
     fn get_merchant_connector_details(
         &self,
-    ) -> Option<api_models::payments::MerchantConnectorDetails> {
+    ) -> Option<common_types::domain::MerchantConnectorAuthDetails> {
         self.merchant_connector_details.clone()
     }
 
@@ -9887,7 +9897,7 @@ impl<F: Clone> OperationSessionGetters<F> for PaymentStatusData<F> {
 
     fn get_merchant_connector_details(
         &self,
-    ) -> Option<api_models::payments::MerchantConnectorDetails> {
+    ) -> Option<common_types::domain::MerchantConnectorAuthDetails> {
         self.merchant_connector_details.clone()
     }
 
@@ -10178,7 +10188,7 @@ impl<F: Clone> OperationSessionGetters<F> for PaymentCaptureData<F> {
 
     fn get_merchant_connector_details(
         &self,
-    ) -> Option<api_models::payments::MerchantConnectorDetails> {
+    ) -> Option<common_types::domain::MerchantConnectorAuthDetails> {
         todo!()
     }
 
@@ -10612,7 +10622,7 @@ impl<F: Clone> OperationSessionGetters<F> for PaymentAttemptListData<F> {
     }
     fn get_merchant_connector_details(
         &self,
-    ) -> Option<api_models::payments::MerchantConnectorDetails> {
+    ) -> Option<common_types::domain::MerchantConnectorAuthDetails> {
         todo!()
     }