+++ crates/router/src/core/payments/helpers.rs
@@ -1602,7 +1602,7 @@ pub async fn get_connector_default(
 #[cfg(feature = "v2")]
 pub async fn get_connector_data_from_request(
     state: &SessionState,
-    req: Option<api_models::payments::MerchantConnectorDetails>,
+    req: Option<common_types::domain::MerchantConnectorAuthDetails>,
 ) -> CustomResult<api::ConnectorData, errors::ApiErrorResponse> {
     let connector = req
         .as_ref()